pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "wstk"
version = "0.1.0"
description = "Portable web search + fetch + extract toolkit for coding/research agents."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "ddgs>=8.0.0",
  "httpx>=0.27.0",
  "markdownify>=0.14.1",
  "readability-lxml>=0.8.1",
  "tldextract>=5.1.2",
]

[project.optional-dependencies]
dev = [
  "pyright>=1.1.390",
  "pytest>=8.0.0",
  "respx>=0.21.0",
  "ruff>=0.6.0",
]

[project.scripts]
wstk = "wstk.cli:main"

[project.urls]
Repository = "https://github.com/tmustier/web-search"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.isort]
known-first-party = ["wstk"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-q"]
